requirements.txt
backports-abc==0.5
bumpversion==0.5.3
Click==7.0
filelock==3.0.10
Jinja2==2.10
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
mkdocs==1.0.4
pluggy==0.8.0
py==1.7.0
ruamel.yaml==0.15.73
dpkt==1.9.2
PyYAML==3.13
singledispatch==3.4.0.3
six==1.11.0
toml==0.10.0
tornado==5.1.1
tox==3.5.3
tox-pyenv==1.1.0
virtualenv==16.1.0
